requirements.txt
contourpy==1.3.2
cycler==0.12.1
fonttools==4.58.1
greenlet==3.2.3
kiwisolver==1.4.8
matplotlib==3.10.3
numpy==2.2.6
packaging==25.0
pandas==2.2.3
pillow==11.2.1
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
seaborn==0.13.2
six==1.17.0
SQLAlchemy==2.0.41
typing_extensions==4.14.0
tzdata==2025.2
uc==2.2.1
uv==0.7.9
streamlit
pg8000